Adjust SSR rendering
